requirements.txt
vpython~=7.6.2
numpy~=1.21.4
pyserial~=3.5